Fixes 2 issues
1. ϵ is declared but never used.
2. Makes BatchNorm notation similar to that of the original paper with similar performance.

### Ubuntu 18.04 GPU K-80

Test
```julia
bn = gpu(BatchNorm(32))
testimg = gpu(randn(Float32, 416, 416, 32, 1))
@time gpu(bn(testimg))
@time gpu(bn(testimg))
```

Old Code
```julia
3.327136 seconds (3.25 M allocations: 167.726 MiB, 1.83% gc time)
0.001245 seconds (155 allocations: 5.391 KiB)
```

New Code
```julia
3.542283 seconds (3.46 M allocations: 178.192 MiB, 2.74% gc time)
0.001185 seconds (149 allocations: 5.281 KiB)
```

### OSX CPU

Test
```julia
using Flux
bn = BatchNorm(32)
testimg = randn(Float32, 416, 416, 32, 1)
@time bn(testimg)
@time bn(testimg)
```

Old Code
```julia
14.939357 seconds (225.12 M allocations: 3.799 GiB, 2.11% gc time)
12.857950 seconds (215.97 M allocations: 3.342 GiB, 1.32% gc time)
```
New Code
```julia
2.701686 seconds (8.37 M allocations: 487.830 MiB, 9.61% gc time)
0.077704 seconds (135 allocations: 63.381 MiB, 24.40% gc time)
```