@KristofferC 2019-02-20
Locally I get

```
julia> @time bn(testimg);
  0.032135 seconds (170 allocations: 63.381 MiB, 15.82% gc time)
```

for old code on cpu.