@staticfloat 2019-02-26
This seems fine to me, but I'm not sure how it would have such a performance difference.  @Sklan what version are you testing when you say "old"?