@Sklan 2019-02-27
@staticfloat I am using Pkg.dev that clones the master branch.
The performance difference is only on the mac cpu and I can't figure out why this works either.