@MikeInnes 2019-03-06
GPU tests pass locally, so I think we can take this. Thanks a lot @Sklan!